1.wiki的id值 获取改用get请求，不再使用wiki_id的url参数

+++ project/views/manage_views.py
@@ -6,6 +6,7 @@ from django.views.generic import DeleteView
 from bugmanage.views import BaseJsonView
 from project.forms import ProjectWikiModelForm
 from project.models import ProjectWikiInfo, ProjectInfo
+from utills.datavalid.datavalid import id_number_vaild
 
 
 class ProjectOverView(View):
@@ -37,32 +38,56 @@ class ProjectFile(View):
         return render(request, ProjectFile.template_name)
 
 
-class ProjectIssue(View):
-    template_name = "project/project_issue.html"
+class ProjectWiki(BaseJsonView):
+    """
+    Attributes
+    -----------
+    template_name：str
+        html模板路径
 
-    def get(self, request, project_id):
-        return render(request, ProjectIssue.template_name)
-
-
-class ProjectWiki(View):
+    Method
+    --------------
+    valid_wiki_id（）验证wiki_id格式是否正确，如果正确并存在，返回wiki
+    get_redirect_url（） 生成project_id wiki主页面或者wiki详情页面的url。通常用于添加，修改，删除后的跳转
+    get() 获取wiki的主页信息或者是wiki的详情
+    """
     template_name = "project/project_wiki.html"
 
+    @staticmethod
+    def valid_wiki_id(request, context, project_id):
+        wiki_id = str(request.GET.get("wiki_id"))
+        if id_number_vaild(wiki_id):
+            wiki = ProjectWikiInfo.objects.filter(id=wiki_id, project_id=project_id).first()
+            if wiki:
+                context["wiki"] = wiki
+                return wiki
+
+    @staticmethod
+    def get_redirect_url(project_id, wiki_id=None):
+        if wiki_id:
+            return "/project/"+str(project_id)+"/"+"wiki/?wiki_id=" + str(wiki_id)
+        return "/project/"+str(project_id)+"/"+"wiki/"
+
     def get(self, request, project_id):
-        return render(request, ProjectWiki.template_name)
-
-
-class ProjectWikiDesc(View):
-    template_name = "project/project_wiki.html"
-
-    def get(self, request, project_id, wiki_id):
-        form = ProjectWikiModelForm(request=request)
-        wiki = ProjectWikiInfo.objects.filter(id=wiki_id).first()
-        if wiki:
-            context = {"wiki": wiki}
+        context = {}
+        if not ProjectWiki.valid_wiki_id(request, context, project_id):
+            pass
         return render(request, ProjectWiki.template_name, context)
 
 
 class ProjectWikiAdd(View):
+    """
+    Attributes
+    -----------
+    template_name：str
+        html模板路径
+
+    Method
+    --------------
+    get() 获取添加wiki的表单页面
+    post（）处理添加wiki的请求
+    """
+
     template_name = "project/project_wiki_add.html"
 
     def get(self, request, project_id):
@@ -79,42 +104,69 @@ class ProjectWikiAdd(View):
             form.save()
         else:
             print(form.errors)
-        return redirect("/project/"+str(project_id)+"/"+"wiki/")
+
+        return redirect(ProjectWiki.get_redirect_url(project_id))
 
 
-class ProjectWikiAlter(View):
+class ProjectWikiAlter(BaseJsonView):
+    """
+    Attributes
+    -----------
+    template_name：str
+        html模板路径
+
+    Method
+    --------------
+    get() 获取添加wiki的修改表单页面，以及返回需要修改wiki的原本信息
+    post（）处理修改wiki的请求
+    """
     template_name = "project/project_wiki_alter.html"
 
-    def get(self, request, project_id, wiki_id):
-        wiki = ProjectWikiInfo.objects.filter(project_id=project_id, id=wiki_id).first()
-        if wiki:
-            form = ProjectWikiModelForm(instance=wiki, request=request)
-        return render(request, ProjectWikiAlter.template_name, {"form": form})
-
-    def post(self, request, project_id, wiki_id):
-        wiki = ProjectWikiInfo.objects.filter(project_id=project_id, id=wiki_id).first()
+    def get(self, request, project_id):
+        context = {}
+        wiki = ProjectWiki.valid_wiki_id(request, context, project_id)
         if not wiki:
-            pass
+            return self.error_response("请求格式不正确")
+        form = ProjectWikiModelForm(instance=wiki, request=request)
+        context["form"] = form
+        return render(request, ProjectWikiAlter.template_name, context)
+
+    def post(self, request, project_id):
+        context = {}
+        wiki = ProjectWiki.valid_wiki_id(request, context, project_id)
+        if not wiki:
+            return self.error_response("请求格式不正确")
         form = ProjectWikiModelForm(data=request.POST, instance=wiki, request=request)
         if form.is_valid():
             form.save()
-            return redirect("/project/" + str(project_id) + "/wiki/" + str(wiki_id) + "/")
+            return redirect(ProjectWiki.get_redirect_url(project_id, wiki.id))
+        return self.error_response("数据验证失败")
 
 
-def project_wiki_delete(request, project_id, wiki_id):
-    wiki = ProjectWikiInfo.objects.filter(project_id=project_id, id=wiki_id)
-    if wiki.exists():
-        wiki.delete()
-    return render(request, ProjectWikiDesc.template_name)
-
-
-
+# 删除wiki
+def project_wiki_delete(request, project_id):
+    wiki_id = request.GET.get("wiki_id")
+    if id_number_vaild(wiki_id):
+        wiki = ProjectWikiInfo.objects.filter(project_id=project_id, id=wiki_id)
+        if wiki.exists():
+            wiki.delete()
+        else:
+            return BaseJsonView.error_response("wiki不存在！")
+    else:
+        BaseJsonView.error_response("请求格式错误！")
+    return render(request, ProjectWiki.template_name)
 
 
 class ProjectWikiTitle(BaseJsonView):
+    """
+    Method
+    --------------
+    get() 根据project_id，获取相关所有标题
+    """
+
     def get(self, request, project_id):
         # 中间件已经进行了项目验证，这里不再做额外验证
-        wiki = ProjectWikiInfo.objects.filter(project_id=project_id).order_by("id").values()
+        wiki = ProjectWikiInfo.objects.filter(project_id=project_id).values()
         context = {"wiki": list(wiki)}
         return self.success_response_data(context)
 
@@ -123,4 +175,4 @@ class ProjectSettings(View):
     template_name = "project/project_settings.html"
 
     def get(self, request, project_id):
-        return render(request, ProjectSettings.template_name)
+        return redirect(ProjectWiki.get_redirect_url(project_id))
